producer/requirements.txt
wheel
kafka-python==2.0.2
python-dotenv==0.19.1

